:truck: Move network test to scripts
